# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Recyclify Bidder Market CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: stage
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
        - name: assignedTo
          in: query
          schema:
            type: integer
        - name: ownerId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: pipelineType
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: sortBy
          in: query
          schema:
            type: string
        - name: sortOrder
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of companies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyListResponse"
    post:
      operationId: createCompany
      tags: [companies]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyInput"
      responses:
        "201":
          description: Company created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/import:
    post:
      operationId: importCompanies
      tags: [companies]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyImportInput"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyImportResult"
  /companies/{id}:
    get:
      operationId: getCompany
      tags: [companies]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
    patch:
      operationId: updateCompany
      tags: [companies]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyUpdate"
      responses:
        "200":
          description: Company updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
    delete:
      operationId: deleteCompany
      tags: [companies]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Company deleted
  /companies/{id}/stage:
    patch:
      operationId: updateCompanyStage
      tags: [companies]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StageUpdate"
      responses:
        "200":
          description: Stage updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/{id}/restore:
    patch:
      operationId: restoreCompany
      tags: [companies]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Company restored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"

  # ── COMPANY BIDS ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    PaginationMeta:
      type: object
      required: [total, page, limit, totalPages]
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer

    # ── AUTH SCHEMAS ────────────────────────────────────────────────────────────
...
    Company:
      type: object
      required: [id, name, stage, status, priority, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        industry:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        gst:
          type: ["string", "null"]
        pan:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        pincode:
          type: ["string", "null"]
        leadSource:
          type: ["string", "null"]
        ownerId:
          type: ["integer", "null"]
        ownerName:
          type: ["string", "null"]
        assignedManagerId:
          type: ["integer", "null"]
        assignedManagerName:
          type: ["string", "null"]
        pipelineType:
          type: string
        stage:
          type: string
        priority:
          type: string
        status:
          type: string
        expectedScrapWeight:
          type: ["number", "null"]
        expectedRevenue:
          type: ["number", "null"]
        expectedPickupDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        deletedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    CompanyListResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Company"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
    CompanyInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        industry:
          type: string
        website:
          type: string
        gst:
          type: string
        pan:
          type: string
        address:
          type: string
        state:
          type: string
        city:
          type: string
        pincode:
          type: string
        leadSource:
          type: string
        ownerId:
          type: integer
        assignedManagerId:
          type: integer
        pipelineType:
          type: string
        stage:
          type: string
        priority:
          type: string
        expectedScrapWeight:
          type: number
        expectedRevenue:
          type: number
        expectedPickupDate:
          type: string
        notes:
          type: string
    CompanyUpdate:
      type: object
      properties:
        name:
          type: string
        industry:
          type: string
        website:
          type: string
        gst:
          type: string
        pan:
          type: string
        address:
          type: string
        state:
          type: string
        city:
          type: string
        pincode:
          type: string
        leadSource:
          type: string
        ownerId:
          type: integer
        assignedManagerId:
          type: integer
        pipelineType:
          type: string
        stage:
          type: string
        priority:
          type: string
        status:
          type: string
        expectedScrapWeight:
          type: number
        expectedRevenue:
          type: number
        expectedPickupDate:
          type: string
        notes:
          type: string
    CompanyImportRow:
      type: object
      properties:
        name:
          type: string
        city:
          type: string
    CompanyImportInput:
      type: object
      required: [rows]
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/CompanyImportRow"
    CompanyImportError:
      type: object
      required: [row, error]
      properties:
        row:
          type: integer
        name:
          type: ["string", "null"]
        error:
          type: string
    CompanyImportResult:
      type: object
      required: [imported, failed, errors]
      properties:
        imported:
          type: integer
        failed:
          type: integer
        errors:
          type: array
          items:
            $ref: "#/components/schemas/CompanyImportError"
    StageUpdate:
      type: object
      required: [stage]
      properties:
        stage:
          type: string

    # ── PIPELINE ────────────────────────────────────────────────────────────────
...