# lib/api-zod/src/generated/types/company.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Recyclify Bidder Market CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface Company {
  id: number;
  name: string;
  /** @nullable */
  industry?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  gst?: string | null;
  /** @nullable */
  pan?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  pincode?: string | null;
  /** @nullable */
  leadSource?: string | null;
  /** @nullable */
  ownerId?: number | null;
  /** @nullable */
  ownerName?: string | null;
  /** @nullable */
  assignedManagerId?: number | null;
  /** @nullable */
  assignedManagerName?: string | null;
  pipelineType?: string;
  stage: string;
  priority: string;
  status: string;
  /** @nullable */
  expectedScrapWeight?: number | null;
  /** @nullable */
  expectedRevenue?: number | null;
  /** @nullable */
  expectedPickupDate?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  deletedAt?: string | null;
  createdAt: string;
  updatedAt?: string;
}