# lib/api-zod/src/generated/types/companyUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Recyclify Bidder Market CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface CompanyUpdate {
  name?: string;
  industry?: string;
  website?: string;
  gst?: string;
  pan?: string;
  address?: string;
  state?: string;
  city?: string;
  pincode?: string;
  leadSource?: string;
  ownerId?: number;
  assignedManagerId?: number;
  pipelineType?: string;
  stage?: string;
  priority?: string;
  status?: string;
  expectedScrapWeight?: number;
  expectedRevenue?: number;
  expectedPickupDate?: string;
  notes?: string;
}